modified backwardWarping to handle double values in images

+++ Sequence/OpticalFlow/backwardWarping.m
@@ -3,7 +3,7 @@ function [ warpIm ] = backwardWarping( im1,im2, flowIm )
 %UNTITLED Summary of this function goes here
 %   Detailed explanation goes here
 [m n c] = size(im1);
-warpIm = zeros(m, n, c,'uint8');
+warpIm = zeros(m, n, c,'double');
 %reconstruct im1 as warpIm from im2 and flow vector
 for x=1:m
         for y= 1:n           
@@ -15,7 +15,7 @@ for x=1:m
             y1 = floor(pWarped(2));
             dx = pWarped(1) - x1;
             dy = pWarped(2) - y1;
-            color = zeros(3,1,'uint8');
+            color = zeros(3,1,'double');
             if( y1 > 0 && x1 > 0 && y1 < n && x1 < m)
                 warpIm(x,y,:)  = im2(x1,y1,:)*(1-dx)*(1-dy)+im2(x1,y1+1,:)*(dx)*(1-dy)+im2(x1+1,y1+1,:)*(dx)*(dy)+im2(x1+1,y1,:)*dy*(1-dx);
